+++ src/problems/cvrp/operators/pd_shift.cpp
@@ -48,8 +48,9 @@ PDShift::PDShift(const Input& input,
     s_gain.cost += _input.vehicles[s_vehicle].fixed_cost();
   }
 
-  assert(_sol_state.route_evals[s_vehicle].duration <=
-         _input.vehicles[s_vehicle].max_travel_time + s_gain.duration);
+  assert(_input.vehicles[s_vehicle].ok_for_travel_time(
+    _sol_state.route_evals[s_vehicle].duration - s_gain.duration));
+
   stored_gain = gain_threshold;
 }