+++ src/structures/vroom/eval.h
@@ -15,15 +15,12 @@ All rights reserved (see LICENSE).
 namespace vroom {
 
 struct Eval {
-  SignedCost cost;
-  SignedDuration duration;
+  Cost cost;
+  Duration duration;
 
   constexpr Eval() : cost(0), duration(0){};
 
   constexpr Eval(Cost cost, Duration duration)
-    : cost(cost), duration(duration){};
-
-  constexpr Eval(SignedCost cost, SignedDuration duration)
     : cost(cost), duration(duration){};
 
   Eval& operator+=(const Eval& rhs) {
@@ -75,8 +72,8 @@ struct Eval {
   }
 };
 
-constexpr Eval NO_EVAL = {std::numeric_limits<SignedCost>::max(), 0};
-constexpr Eval NO_GAIN = {std::numeric_limits<SignedCost>::min(), 0};
+constexpr Eval NO_EVAL = {std::numeric_limits<Cost>::max(), 0};
+constexpr Eval NO_GAIN = {std::numeric_limits<Cost>::min(), 0};
 
 } // namespace vroom