+++ src/structures/vroom/input/input.h
@@ -40,6 +40,7 @@ private:
   bool _no_addition_yet;
   bool _has_skills;
   bool _has_TW;
+  bool _has_all_coordinates;
   bool _has_custom_location_index;
   bool _has_initial_routes;
   bool _homogeneous_locations;
@@ -48,8 +49,8 @@ private:
   bool _geometry;
   bool _has_jobs;
   bool _has_shipments;
-  std::unordered_map<std::string, Matrix<Duration>> _durations_matrices;
-  std::unordered_map<std::string, Matrix<Duration>> _costs_matrices;
+  std::unordered_map<std::string, Matrix<UserDuration>> _durations_matrices;
+  std::unordered_map<std::string, Matrix<UserCost>> _costs_matrices;
   Cost _cost_upper_bound;
   std::vector<Location> _locations;
   std::unordered_map<Location, Index> _locations_to_index;
@@ -70,7 +71,7 @@ private:
 
   void check_job(Job& job);
 
-  Cost check_cost_bound(const Matrix<Cost>& matrix) const;
+  UserCost check_cost_bound(const Matrix<UserCost>& matrix) const;
 
   void set_skills_compatibility();
   void set_extra_compatibility();
@@ -103,8 +104,9 @@ public:
 
   void add_vehicle(const Vehicle& vehicle);
 
-  void set_durations_matrix(const std::string& profile, Matrix<Duration>&& m);
-  void set_costs_matrix(const std::string& profile, Matrix<Cost>&& m);
+  void set_durations_matrix(const std::string& profile,
+                            Matrix<UserDuration>&& m);
+  void set_costs_matrix(const std::string& profile, Matrix<UserCost>&& m);
 
   const Amount& zero_amount() const {
     return _zero;