+++ src/structures/vroom/vehicle.cpp
@@ -11,6 +11,7 @@ All rights reserved (see LICENSE).
 
 #include "structures/vroom/vehicle.h"
 #include "utils/exception.h"
+#include "utils/helpers.h"
 
 namespace vroom {
 
@@ -26,7 +27,7 @@ Vehicle::Vehicle(Id id,
                  const VehicleCosts& costs,
                  double speed_factor,
                  const size_t max_tasks,
-                 const Duration max_travel_time,
+                 const std::optional<UserDuration>& max_travel_time,
                  const std::vector<VehicleStep>& input_steps)
   : id(id),
     start(start),
@@ -40,7 +41,9 @@ Vehicle::Vehicle(Id id,
     costs(costs),
     cost_wrapper(speed_factor),
     max_tasks(max_tasks),
-    max_travel_time(max_travel_time) {
+    max_travel_time(max_travel_time.has_value()
+                      ? utils::scale_from_user_duration(max_travel_time.value())
+                      : std::numeric_limits<Duration>::max()) {
   if (!static_cast<bool>(start) and !static_cast<bool>(end)) {
     throw InputException("No start or end specified for vehicle " +
                          std::to_string(id) + '.');
@@ -111,8 +114,12 @@ bool Vehicle::has_same_locations(const Vehicle& other) const {
 bool Vehicle::has_same_profile(const Vehicle& other) const {
   return (this->profile == other.profile) and
          (this->costs.fixed == other.costs.fixed) and
-         (this->cost_wrapper.discrete_duration_factor ==
-          other.cost_wrapper.discrete_duration_factor);
+         (this->cost_wrapper.get_discrete_duration_factor() ==
+          other.cost_wrapper.get_discrete_duration_factor());
+}
+
+bool Vehicle::cost_is_duration() const {
+  return cost_wrapper.cost_is_duration();
 }
 
 Duration Vehicle::available_duration() const {