+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Refactor `TSP` cost functions (#812)
 - CI builds now use clang++ 14 and g++ 11 on Ubuntu 22.04 (#816)
 - Refactor `CVRP::solve` and `VRPTW::solve` functions (#818)
+- Refactor `CostWrapper` (#828)
 
 ### Fixed
 
@@ -30,6 +31,8 @@
 - Update `ssl_send_and_receive` to throw RoutingExceptions (#770)
 - Timeout not observed with multiple long heuristics per thread (#792)
 - Wrong validity check range in `vrptw::MixedExchange` (#821)
+- Underflow in insertion regrets (#831)
+- Crash with missing location coordinates and only `costs` custom matrix (#826)
 
 ## [v1.12.0] - 2022-05-31